Add files via upload
A small text to speech GUI app to track files coming into a folder over a set interval of time. 

Use case could be for network folders where files arrive and you do not want to manually check each time. You'd need to modify the py file to specify the folder paths and the file extensions you'd like to track. Currently has functionality to track two folders, more can be added. The app will give you a speech alert of the count of files in the folder.